+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# v11.4.3
+
+## Enhancements
+- [#220] Default Character, Robot and Settlement actor types to have linked actor data by default
+
+## Bugfixes
+- [#221] Sort owned items on Creature/NPC sheets
+
+## Chores
+- [#218] Merged i18n changes from Crowdin
+
+---
+
 # v11.4.2
 
 ## Enhancements